ai_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
ultralytics==8.1.0
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
pika==1.3.2
